feat(comet): custom fonts on the Compose backend — Jetchat renders Montserrat + Karla
Extend the vendored facade with the one clean interop binding it lacked and wire it
through, so custom fonts (and, next, icon fonts) work on the node backend:

- Facade: bind androidx.compose.ui.text.font.AndroidTypeface_androidKt.FontFamily(Typeface)
  (found the exact Kt class via javap on ui-text-android.aar) + FontFamily.FromTypeface().
- ComposeFontRegistry: app registers a base Typeface per family; Resolve(family, weight)
  derives a weighted Typeface (Typeface.Create, API 28+) and wraps it as a Compose
  FontFamily, cached, degrading to the default font if the JNI path ever fails (no crash).
- ComposeTextNode: applies the resolved FontFamily (the typeface carries the weight, so it
  skips FontWeight then) and measures with the same Typeface so metrics match the render.
- App loads Montserrat + Karla (bundled variable TTFs) from assets and registers them.

Also fixes a real latent bug: Text_FontWeight is emitted as From(int) but was read via
AsDouble on BOTH backends (interpreting 600 as an IEEE double bit-pattern → ~0 → Regular).
Now read via AsInt, so weights actually apply (Montserrat@600 titles, Karla@700/@400).

Verified on Pixel 5: names/title in Montserrat SemiBold, body/timestamps in Karla. 44
backend tests pass.

+++ src/Comet/src/Comet/Platform/SwiftUI/SwiftUINode.cs
@@ -71,7 +71,7 @@ namespace Comet.Platform.SwiftUI
 			else if (id == PropertyIds.Text_FontSize)
 				CometSwiftUIHost.SetDouble(_native, "fontsize", value.AsDouble);
 			else if (id == PropertyIds.Text_FontWeight)
-				CometSwiftUIHost.SetDouble(_native, "fontweight", value.AsDouble);
+				CometSwiftUIHost.SetDouble(_native, "fontweight", value.AsInt);   // Int-kind, not Double
 			else if (id == PropertyIds.BackgroundColor && value.AsColor is { } c)
 				CometSwiftUIHost.SetColor(_native, "background", ToArgb(c));
 			else if (id == PropertyIds.Padding && value.AsObject is Microsoft.Maui.Thickness t)
